変更

+++ db/migrate/20201216051135_create_evals.rb
@@ -1,7 +1,7 @@
 class CreateEvals < ActiveRecord::Migration[5.2]
   def change
     create_table :evals do |t|
-      t.integer :ceval
+      t.float :ceval
       t.string :cmt
       t.integer :user_id
       t.integer :store_id
